+++ views/admin/index.pug
@@ -1,0 +1,9 @@
+extends layout
+
+block content
+  h1 Administration
+  ul.menu
+    li
+      a(href="/admin/create") New post
+    li
+      a(href="/admin/settings") Settings
