+++ views/create/full.pug
@@ -1,4 +1,4 @@
-form(action="/new", method="post")
+form(action="/admin/new", method="post")
   p If you want to be more elaborate, feel free.
   input(type="hidden", name="type", value="full")
   input#full-title(type="text", placeholder="title", name="title")
