+++ views/create/link.pug
@@ -1,4 +1,4 @@
-form(action="/new", method="post")
+form(action="/admin/new", method="post")
   p Link post
   input(type="hidden", name="type", value="link")
   input#link-title(type="text", placeholder="title", name="title")
