+++ views/create/note.pug
@@ -1,4 +1,4 @@
-form(action="/new", method="post")
+form(action="/admin/new", method="post")
   p Just write a note, it can be anyting, and doesn't have a title
   input(type="hidden", name="type", value="note")
   textarea#link-content(name="content", cols="30", rows="10")
